requirements.txt
beautifulsoup4==4.14.3
certifi==2026.2.25
chardet==3.0.4
charset-normalizer==3.4.6
click==8.1.8
cuda-bindings==13.2.0
cuda-pathfinder==1.5.0
cuda-toolkit==13.0.2
decorator==4.4.2
deep-translator==1.11.4
filelock==3.25.2
fsspec==2026.3.0
googletrans==4.0.0rc1
gTTS==2.5.4
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
ImageIO==2.37.3
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
legacy-cgi==2.6.4
llvmlite==0.47.0
MarkupSafe==3.0.3
more-itertools==10.8.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.6.1
numba==0.65.0
numpy==2.4.4
nvidia-cublas==13.1.0.3
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
openai-whisper==20250625
pillow==11.3.0
proglog==0.1.12
pydub==0.25.1
python-dotenv==1.2.2
regex==2026.3.32
requests==2.33.1
rfc3986==1.5.0
setuptools==81.0.0
sniffio==1.3.1
soupsieve==2.8.3
sympy==1.14.0
tiktoken==0.12.0
torch==2.11.0
tqdm==4.67.3
triton==3.6.0
typing_extensions==4.15.0
urllib3==2.6.3
yt-dlp==2026.3.17
